requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-environ==0.10.0
psycopg2-binary==2.9.6
django-choices==2.0.0